Move `<GameOverOverlay/>` and `<PlayAgainButton/>` to own file
